Refactor: Organize JavaScript files into structured folders
- Created 'components' folder for reusable UI components
  - Moved 'carouselSlider.js' and 'countdown.js' to 'components'

- Created 'services' folder for service-related files
  - Moved 'jwtToken.js' to 'services'

- Created 'utils' folder for utility functions and helpers
  - Moved 'itemUtils.js', 'utils.js', and 'numberInputRestrictor.js' to 'utils'

- Created 'handlers' folder for event handling logic
  - Moved 'eventHandlers.js', 'handleTextCharInput.js', and 'imageRotationHandler.js' to 'handlers'

- Created 'modules' folder for modular application features
  - Moved 'cart.js',  'reviews.js', and 'add-review.js' to 'modules'

- Created 'pages' folder for page-specific scripts
  - Moved 'main.js', 'view-item.js', 'orders.js', and 'invoice.js' to 'pages'

- Created 'builders' folder for build logic
  - Moved 'builder.js' to 'builders'

- Created 'messages' folder for managing user messages
  - Moved 'messages.js' to 'messages'

The aim is to improve maintainability and readability of the codebase by logically grouping related files.
