+++ static/js/components/carouselSlider.js
