+++ static/js/handlers/handleTextCharInput.js
