+++ static/js/handlers/imageRotationHandler.js
