+++ static/js/pages/invoice.js
@@ -1,5 +1,5 @@
 import getItemByID from "./itemUtils.js";
-import orders from "../../order.js";
+import orders from "../../../order.js";
 import { getItemFromLocalStorage, getFormattedCurrentDate } from "./utils.js";
 
 const REQUIRED_ELEMENTS_COUNT = 3;
