+++ static/js/utils/utils.js
