+++ templates/account/orders/orders.html
@@ -304,7 +304,7 @@
                                         <button class="buy-again-btn order-btn">Buy again</button>
                                         <button class="view-item-btn order-btn" data-order-id="1">View item</button>
                                         <button class="view-invoice-btn order-btn" data-order-id="1">View Invoice</button>
-                                        <button class="leave-feedback-btn order-btn">Leave Feedback</button>
+                                        <button class="leave-feedback-btn order-btn" data-order-id="1">Leave Feedback</button>
                                     </div>
                                 </div>
